refactor(cli): split serve server assembly
