+++ packages/cli/src/serve/server/device-flow-registry.ts
@@ -1,0 +1,99 @@
+/**
+ * @license
+ * Copyright 2025 Qwen Team
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import type { Application } from 'express';
+import { writeStderrLine } from '../../utils/stdioHelpers.js';
+import type { AcpSessionBridge } from '../acp-session-bridge.js';
+import {
+  DeviceFlowRegistry,
+  setDeviceFlowRegistry,
+  type DeviceFlowEventSink,
+  type DeviceFlowProvider,
+  type DeviceFlowProviderId,
+} from '../auth/device-flow.js';
+import { QwenOAuthDeviceFlowProvider } from '../auth/qwen-device-flow-provider.js';
+
+interface SetupDeviceFlowRegistryDeps {
+  app: Application;
+  bridge: AcpSessionBridge;
+  registry?: DeviceFlowRegistry;
+  providers?: DeviceFlowProvider[];
+}
+
+export interface ServeDeviceFlowRuntime {
+  deviceFlowRegistry: DeviceFlowRegistry;
+  getSupportedDeviceFlowProviders: () => DeviceFlowProviderId[];
+}
+
+export function setupDeviceFlowRegistry(
+  deps: SetupDeviceFlowRegistryDeps,
+): ServeDeviceFlowRuntime {
+  const deviceFlowProviderMap = new Map<
+    DeviceFlowProviderId,
+    DeviceFlowProvider
+  >();
+  for (const provider of deps.providers ?? []) {
+    deviceFlowProviderMap.set(provider.providerId, provider);
+  }
+  if (!deviceFlowProviderMap.has('qwen-oauth')) {
+    deviceFlowProviderMap.set('qwen-oauth', new QwenOAuthDeviceFlowProvider());
+  }
+
+  const deviceFlowEventSink: DeviceFlowEventSink = {
+    publish(emission, originatorClientId) {
+      deps.bridge.publishWorkspaceEvent({
+        type: `auth_device_flow_${emission.type}`,
+        data: emission.data,
+        ...(originatorClientId ? { originatorClientId } : {}),
+      });
+    },
+  };
+  const deviceFlowRegistry =
+    deps.registry ??
+    new DeviceFlowRegistry({
+      events: deviceFlowEventSink,
+      audit: {
+        record(line) {
+          // Structured stderr breadcrumb; deviceFlowId truncated to first
+          // 8 chars so log skimmers can follow a flow without retaining
+          // full uuids.
+          const id = line.deviceFlowId.slice(0, 8);
+          const parts = [
+            `[serve] auth.device-flow:`,
+            `provider=${line.providerId}`,
+            `deviceFlowId=${id}...`,
+            line.clientId ? `clientId=${line.clientId}` : 'clientId=-',
+            `status=${line.status}`,
+          ];
+          if (line.errorKind) parts.push(`errorKind=${line.errorKind}`);
+          if (line.expiresInMs !== undefined) {
+            parts.push(`expiresInMs=${Math.max(0, line.expiresInMs)}`);
+          }
+          // Include `line.hint` for operator-only breadcrumbs that aren't
+          // surfaced over SSE. Bound at 1 KiB.
+          if (line.hint) {
+            const STDERR_HINT_MAX = 1_024;
+            const hint =
+              line.hint.length > STDERR_HINT_MAX
+                ? `${line.hint.slice(0, STDERR_HINT_MAX)}…[+${line.hint.length - STDERR_HINT_MAX} bytes truncated]`
+                : line.hint;
+            // Quote the hint so multi-word values stay parseable.
+            parts.push(`hint=${JSON.stringify(hint)}`);
+          }
+          writeStderrLine(parts.join(' '));
+        },
+      },
+      resolveProvider: (providerId) => deviceFlowProviderMap.get(providerId),
+    });
+
+  setDeviceFlowRegistry(deps.app, deviceFlowRegistry);
+
+  return {
+    deviceFlowRegistry,
+    getSupportedDeviceFlowProviders: () =>
+      Array.from(deviceFlowProviderMap.keys()),
+  };
+}
